+++ core/src/baseTypes.ts
@@ -1,0 +1,10 @@
+import { Prefix } from "./prefix";
+import { QuantityKind } from "./quantityKind";
+import { Unit } from "./unit";
+
+export type UnitOrExponent = Unit | number;
+export type ExponentUnitPairs = UnitOrExponent[];
+
+export interface SupportsEquals<Type> {
+  equals(other?: Type): boolean;
+}
