+++ core/src/langString.ts
@@ -1,0 +1,23 @@
+import { SupportsEquals } from "./baseTypes";
+
+export class LangString implements SupportsEquals<LangString> {
+  readonly text: string;
+  readonly languageTag?: string;
+
+  constructor(text: string, languageTag?: string) {
+    this.text = text;
+    this.languageTag = languageTag;
+  }
+
+  equals(other?: LangString): boolean {
+    return (
+      !!other &&
+      this.text === other.text &&
+      this.languageTag === other.languageTag
+    );
+  }
+
+  toString(): string {
+    return this.text + (this.languageTag ? `@${this.languageTag}` : "");
+  }
+}
